case: Hack around incorrect motherboard height
The IO cutout is too high and the motherboard IO area will not fit properly.
We need longer standoffs which is impractical.

The issue is the stiffening ribs should not have been added to the height.
Removing it now will cause an avalanche and the whole slicing will have to be redone.

We also need longer standoffs than the 6.5 I am planning to use to allow the end of PCIE brackets
to sit. It should be 9.5 or we curve a slot into the bottom for the ends which is too much work.
I am however using 12 by stacking 2 x 6mm since it is what I have.

The whole plan was to make the motherboard low enough to allow the fan hinge but it is what it is now.

I will just adjust the part since it is the easiest way and design a new fan mount.